FeaturesProvider/requirements.txt
fastapi
uvicorn
httpx
requests
python-dotenv
loguru